Various additions.
Making a commit here since I like the look of the 'Relation Instances' section as it is, but I want to make it smaller to fit more content in. This commit should hopefully mean I can recover this version in the future if needed.

Additionally, I made some major changes, notably:
    - changing \CheatsheetEntryFrame into an environment, and
    - preferring to use scopes for locally-useful command definitions rather than defining placeholders for commands and redefining it throughout the document.
All code should be migrated to these new structures in its own commit.